+++ CHANGES.rst
@@ -8,6 +8,7 @@ Changelog
 * **Breaking change**: Changed CLI command format. Before: ``darq some_project.darq_app.darq``. Now: ``darq -A some_project.darq_app.darq worker``
 * **Breaking change**: Scheduler (cron jobs) now run's seperate from worker (see ``darq scheduler`` command)
 * **Breaking change**: Changed some function signatures (rename arguments)
+* **Breaking change**: Remove ``redis_pool`` param from ``Darq`` app
 
 0.7.2 (2020-06-18)
 ...................